nvme: move id commands into a plugin
There are too many top-level commands cluttering the space. Move the
id commands into a plugin, e.g.,

  nvme id-ctrl -> nvme id ctrl
