+++ src/nvme-builtin.h
@@ -10,24 +10,26 @@
 COMMAND_LIST(
 	ENTRY("list", "List all NVMe devices and namespaces on machine", list)
 	ENTRY("list-subsys", "List nvme subsystems", list_subsys)
-	ENTRY("id-ctrl", "Send NVMe Identify Controller", id_ctrl)
-	ENTRY("id-ns", "Send NVMe Identify Namespace, display structure", id_ns)
-	ENTRY("id-ns-granularity", "Send NVMe Identify Namespace Granularity List, display structure", id_ns_granularity)
-	ENTRY("id-ns-lba-format", "Send NVMe Identify Namespace for the specified LBA Format index, display structure", id_ns_lba_format)
-	ENTRY("list-ns", "Send NVMe Identify List, display structure", list_ns)
-	ENTRY("list-ctrl", "Send NVMe Identify Controller List, display structure", list_ctrl)
-	ENTRY("nvm-id-ctrl", "Send NVMe Identify Controller NVM Command Set, display structure", nvm_id_ctrl)
-	ENTRY("nvm-id-ns", "Send NVMe Identify Namespace NVM Command Set, display structure", nvm_id_ns)
-	ENTRY("nvm-id-ns-lba-format", "Send NVMe Identify Namespace NVM Command Set for the specified LBA Format index, display structure", nvm_id_ns_lba_format)
-	ENTRY("primary-ctrl-caps", "Send NVMe Identify Primary Controller Capabilities", primary_ctrl_caps)
-	ENTRY("list-secondary", "List Secondary Controllers associated with a Primary Controller", list_secondary_ctrl)
-	ENTRY("cmdset-ind-id-ns", "I/O Command Set Independent Identify Namespace", cmd_set_independent_id_ns)
-	ENTRY("ns-descs", "Send NVMe Namespace Descriptor List, display structure", ns_descs)
-	ENTRY("id-nvmset", "Send NVMe Identify NVM Set List, display structure", id_nvmset)
-	ENTRY("id-uuid", "Send NVMe Identify UUID List, display structure", id_uuid)
-	ENTRY("id-iocs", "Send NVMe Identify I/O Command Set, display structure", id_iocs)
-	ENTRY("id-domain", "Send NVMe Identify Domain List, display structure", id_domain)
-	ENTRY("list-endgrp", "Send NVMe Identify Endurance Group List, display structure", id_endurance_grp_list)
+#ifdef CONFIG_DEPRECATED_CMDS
+	ENTRY_DEPRECATED("id-ctrl", "Send NVMe Identify Controller (deprecated, use 'nvme id ctrl')", id_ctrl)
+	ENTRY_DEPRECATED("id-ns", "Send NVMe Identify Namespace, display structure (deprecated, use 'nvme id ns')", id_ns)
+	ENTRY_DEPRECATED("id-ns-granularity", "Send NVMe Identify Namespace Granularity List, display structure (deprecated, use 'nvme id ns-granularity')", id_ns_granularity)
+	ENTRY_DEPRECATED("id-ns-lba-format", "Send NVMe Identify Namespace for the specified LBA Format index, display structure (deprecated, use 'nvme id ns-lba-format')", id_ns_lba_format)
+	ENTRY_DEPRECATED("list-ns", "Send NVMe Identify List, display structure (deprecated, use 'nvme id ns-list')", list_ns)
+	ENTRY_DEPRECATED("list-ctrl", "Send NVMe Identify Controller List, display structure (deprecated, use 'nvme id ctrl-list')", list_ctrl)
+	ENTRY_DEPRECATED("nvm-id-ctrl", "Send NVMe Identify Controller NVM Command Set, display structure (deprecated, use 'nvme id nvm-ctrl')", nvm_id_ctrl)
+	ENTRY_DEPRECATED("nvm-id-ns", "Send NVMe Identify Namespace NVM Command Set, display structure (deprecated, use 'nvme id nvm-ns')", nvm_id_ns)
+	ENTRY_DEPRECATED("nvm-id-ns-lba-format", "Send NVMe Identify Namespace NVM Command Set for the specified LBA Format index, display structure (deprecated, use 'nvme id nvm-ns-lba-format')", nvm_id_ns_lba_format)
+	ENTRY_DEPRECATED("primary-ctrl-caps", "Send NVMe Identify Primary Controller Capabilities (deprecated, use 'nvme id primary-ctrl-caps')", primary_ctrl_caps)
+	ENTRY_DEPRECATED("list-secondary", "List Secondary Controllers associated with a Primary Controller (deprecated, use 'nvme id secondary-ctrl-list')", list_secondary_ctrl)
+	ENTRY_DEPRECATED("cmdset-ind-id-ns", "I/O Command Set Independent Identify Namespace (deprecated, use 'nvme id ns-ind')", cmd_set_independent_id_ns)
+	ENTRY_DEPRECATED("ns-descs", "Send NVMe Namespace Descriptor List, display structure (deprecated, use 'nvme id ns-descs')", ns_descs)
+	ENTRY_DEPRECATED("id-nvmset", "Send NVMe Identify NVM Set List, display structure (deprecated, use 'nvme id nvmset')", id_nvmset)
+	ENTRY_DEPRECATED("id-uuid", "Send NVMe Identify UUID List, display structure (deprecated, use 'nvme id uuid')", id_uuid)
+	ENTRY_DEPRECATED("id-iocs", "Send NVMe Identify I/O Command Set, display structure (deprecated, use 'nvme id iocs')", id_iocs)
+	ENTRY_DEPRECATED("id-domain", "Send NVMe Identify Domain List, display structure (deprecated, use 'nvme id domain')", id_domain)
+	ENTRY_DEPRECATED("list-endgrp", "Send NVMe Identify Endurance Group List, display structure (deprecated, use 'nvme id endgrp-list')", id_endurance_grp_list)
+#endif /* CONFIG_DEPRECATED_CMDS */
 	ENTRY("create-ns", "Creates a namespace with the provided parameters", create_ns)
 	ENTRY("delete-ns", "Deletes a namespace from the controller", delete_ns)
 	ENTRY("attach-ns", "Attaches a namespace to requested controller(s)", attach_ns)
